+++ lib/elastic_apm/spies/sidekiq.rb
@@ -14,7 +14,8 @@ module ElasticAPM
         def call(_worker, job, queue)
           name = SidekiqSpy.name_for(job)
           transaction = ElasticAPM.start_transaction(name, 'Sidekiq')
-          ElasticAPM.set_tag(:queue, queue)
+          # TODO: Remove #to_s when #set_tag is removed in v3.0
+          ElasticAPM.set_label(:queue, queue.to_s)
 
           yield