+++ lib/elastic_apm/transport/serializers/context_serializer.rb
@@ -10,7 +10,7 @@ module ElasticAPM
 
           {
             custom: context.custom,
-            tags: keyword_object(context.tags),
+            tags: mixed_object(context.labels),
             request: build_request(context.request),
             response: build_response(context.response),
             user: build_user(context.user)