+++ spec/elastic_apm/error_builder_spec.rb
@@ -31,7 +31,7 @@ module ElasticAPM
               ElasticAPM.build_context rack_env: env, for_type: :transaction
 
             ElasticAPM.with_transaction context: context do |txn|
-              ElasticAPM.set_tag(:my_tag, '123')
+              ElasticAPM.set_label(:my_tag, '123')
               ElasticAPM.set_custom_context(all_the_other_things: 'blah blah')
               ElasticAPM.set_user(Struct.new(:id).new(321))
               ElasticAPM.report actual_exception
@@ -44,7 +44,7 @@ module ElasticAPM
         expect(error.transaction).to eq(sampled: true, type: 'custom')
         expect(error.transaction_id).to eq transaction.id
         expect(error.trace_id).to eq transaction.trace_id
-        expect(error.context.tags).to match(my_tag: '123', more: 'totes')
+        expect(error.context.labels).to match(my_tag: '123', more: 'totes')
         expect(error.context.custom)
           .to match(all_the_other_things: 'blah blah')
       end