+++ spec/elastic_apm_spec.rb
@@ -176,7 +176,7 @@ RSpec.describe ElasticAPM do
       should delegate :report_message,
         to: agent, args: ['NOT OK', { backtrace: Array, context: nil }]
     end
-    it { should delegate :set_tag, to: agent, args: [nil, nil] }
+    it { should delegate :set_label, to: agent, args: [nil, nil] }
     it { should delegate :set_custom_context, to: agent, args: [nil] }
     it { should delegate :set_user, to: agent, args: [nil] }
 
@@ -264,6 +264,21 @@ RSpec.describe ElasticAPM do
         ensure
           ElasticAPM.stop
         end
+      end
+    end
+
+    describe '.set_tag' do
+      it 'redirects to set_label with string value' do
+        expect(ElasticAPM).to receive(:warn).with(/DEPRECATED/).exactly(3).times
+
+        expect(ElasticAPM).to receive(:set_label).with(:a_string, 'a')
+        ElasticAPM.set_tag(:a_string, 'a')
+
+        expect(ElasticAPM).to receive(:set_label).with(:a_number, '1')
+        ElasticAPM.set_tag(:a_number, 1)
+
+        expect(ElasticAPM).to receive(:set_label).with(:a_boolean, 'true')
+        ElasticAPM.set_tag(:a_boolean, true)
       end
     end
   end